Update delay in increment_loading()
Testing revealed that time.sleep(0) delays more than avoiding the block with an if-statement. On repl.it, 1 000 000 calls to time.sleep() delayed the program by around 7 seconds, meaning each call actually delays by 7 microseconds. For 10-15 calls, the delay is negligible, but if this program is called frequently, the time adds up.